requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
asyncio==3.4.3
attrs==24.2.0
certifi==2024.8.30
Deprecated==1.2.15
frozenlist==1.5.0
idna==3.10
img2pdf==0.5.1
lxml==5.3.0
magic-filter==1.0.12
multidict==6.1.0
packaging==24.2
pikepdf==9.4.2
pillow==11.0.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
typing_extensions==4.12.2
wrapt==1.17.0
yarl==1.18.3
